feat(data): unzip and file detection in script

+++ data/gun-detection.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
 mkdir sources
-curl -L -o ./sources/gun-detection.zip https://www.kaggle.com/api/v1/datasets/download/ugorjiir/gun-detection
+if [ -f ./sources/gun-detection.zip ]; then
+    echo "File already exists. Skipping download."
+else
+    echo "Downloading gun detection dataset..."
+    curl -L -o ./sources/gun-detection.zip https://www.kaggle.com/api/v1/datasets/download/ugorjiir/gun-detection
+fi
+unzip ./sources/gun-detection.zip -d ./sources
